requirements.txt
attrs
aws-cdk.assets
aws-cdk.aws-acmpca
aws-cdk.aws-apigateway
aws-cdk.aws-apigatewayv2
aws-cdk.aws-apigatewayv2-integrations
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling
aws-cdk.aws-autoscaling-common
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudwatch
aws-cdk.aws-cloudwatch-actions
aws-cdk.aws-codeguruprofiler
aws-cdk.aws-codestarnotifications
aws-cdk.aws-cognito
aws-cdk.aws-dynamodb
aws-cdk.aws-ec2
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-efs
aws-cdk.aws-elasticloadbalancing
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-iam
aws-cdk.aws-kinesis
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-lambda-destinations
aws-cdk.aws-logs
aws-cdk.aws-route53
aws-cdk.aws-rds
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-servicediscovery
aws-cdk.aws-signer
aws-cdk.aws-sns
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.aws-stepfunctions
aws-cdk.aws-stepfunctions-tasks
aws-cdk.aws-wafv2
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.region-info
black
boto3
botocore
cattrs
click
colorama
constructs
jmespath
jsii
MacFSEvents
mypy-extensions
nose
pathspec
platformdirs
publication
python-dateutil
python-termstyle
s3transfer
six
sniffer
tomli
typing_extensions
urllib3
